Confirmation Expected Text moved to FeeTarget Converter
